Loader sizes
Added support for 5 m and 7 m loaders in anticipation for the coming update.
Added the ability to set a minimum shell length. The minimum length is EXCLUSIVE to more easily conform to loader sizes: a minimum of 4,000 and maximum of 5,000 will yield shells which would fit into a 5 m loader but not a 4 m loader.