Removed Console Project
I originally wrote this optimizer as a Console App, with the UI as a separate Visual Studio Project. In preparation for significant changes in the near future, I have removed the old console app and integrated all its classes into the APSCalcUI project. This frees me from having to update the console application to incorporate new changes, and reduces the zipped folder size by precisely 51,462 bytes (0.154%).