Added slopes and wedges
Added the ability to include slopes and wedges to the target armor configuration.

Note that all slopes are assumed to
be built with the airgap facing outwards, away from the center of the craft.
Additionally, each sloped block has a "shallow" and a "steep" option, for
shells hitting parallel or perpendicular to the slope. Shallow angles are
closer to parallel with the surface of the block, and thus will have a more
severe kinetic damage reduction and a greater ricochet chance.